Beginning to test AntiCachingPlugin

+++ src/plugins/knapsack_solver.hpp
@@ -6,6 +6,6 @@
 
 class KnapsackSolver {
  public:
-  static std::vector<size_t> solve(const uint64_t memory_budget, const std::vector<uint64_t> values, const std::vector<uint64_t> costs);
+  static std::vector<size_t> solve(const uint64_t memory_budget, const std::vector<float> values, const std::vector<uint64_t> costs);
 };
